Refactor best overall solution calculation to include unsplittable bids. The solution calculator now operates based on groups of bids to determine the best overall solution. Each splittable bid from a supplier is given its own group, and another group containing all bids from the supplier is added if the supplier has any non-splittable bids.
To find the best solution, the calculator then considers all combinations of these groups to find the ones that cover all products in the auction uniquely and then selects the set with the lowest total price.

The advantage of this group-oriented structure is that the calculation can remain constant even if the determination of what "splittable" means for bids. IE, if the semantics change such that nonsplittable bids only need one additional product to be paired with them, then the only thing that needs to change is the generation of the groups. The calculation from them can remain the same.