Support full path of creating a term (at least forward-fixed) auction. From selecting a "Forward Fixed" auction in the form all the way up to (but not fully including) viewing the new auction on the show page, Term auctions are now able to be created. The controller is mostly able to stay agnostic with the exception of one cast of `vessels` vs `auction_vessel_fuels` for parameters.
The tests on the show page are currently failing because `suppliers` is not being loaded properly on these auctions. This has to do with how we load payloads for the show page and that `Auctions.list_auctions` doesn't know how to distinguish between auction types. I think the best path forward will be to have `list_auctions` query with a `type: "spot"` condition to load regular Auctions, then query again with `type: ["forward_fixed", "formula_related"]` to get TermAuctions.

The types should also be moved into constants instead of being given as literal lists everywhere.