This branch fixes the Spot Auction creation form not rendering the general error tag when creating an auction without a fuel.
This error was happening because that particular flag was only checking
if there were errors in the general auction changeset and not the
auction_vessel_fuel changesets that might be nested within it.

Also mix formatted branch and fixed some red in the formula related show
test. The AuctionClosed email was looking for a broker entity because
the test auction was traded bid, but one was not being inserted into the
setup. This might be reoccurring in other places.